+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Added a screen shake effect on collisions
+
 - The viewport is now independent from browser resolution and pixel density.
   It scales itself to contain all sensor readings